refactoring location and also adding simple test
