+++ src/Bet.Extensions.ML.Spam/ServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using Bet.Extensions.ML.ModelBuilder;
+using Bet.Extensions.ML.ModelStorageProviders;
 using Bet.Extensions.ML.Spam;
 using Bet.Extensions.ML.Spam.Models;
 
@@ -9,6 +10,11 @@ namespace Microsoft.Extensions.DependencyInjection
 {
     public static class ServiceCollectionExtensions
     {
+        /// <summary>
+        /// Adds ML.NET based Spam detection model builder with <see cref="FileModelStorageProvider"/> as storage.
+        /// </summary>
+        /// <param name="services"></param>
+        /// <returns></returns>
         public static IServiceCollection AddSpamDetectionModelBuilder(this IServiceCollection services)
         {
             services.TryAddSingleton(new MLContext());
