+++ src/Bet.Extensions.ML/Bet.Extensions.ML.csproj
@@ -21,9 +21,4 @@
     <ProjectReference Include="..\Bet.Extensions\Bet.Extensions.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="ModelStorageProviders\" />
-  </ItemGroup>
-
-
 </Project>
