+++ src/Bet.Extensions.ML/ModelStorageProviders/InMemoryModelStorageProvider.cs
@@ -10,7 +10,7 @@ using Microsoft.Extensions.Primitives;
 
 using Newtonsoft.Json;
 
-namespace Bet.Extensions.ML.ModelBuilder
+namespace Bet.Extensions.ML.ModelStorageProviders
 {
     public class InMemoryModelStorageProvider : IModelStorageProvider
     {
